refactor(map): relocate WaypointsLayer to map/layers/ (#4047 Phase 7 WP4)
Clean move of the already-shared PerSourceWaypoints/WaypointsLayer from
MapAnalysis/layers/ into map/layers/ (it has no MapAnalysis-specific
dependency besides its own default export's useMapAnalysisCtx read).
Update both direct importers (DashboardWaypoints, MapAnalysisCanvas) —
NodesTab picks it up transitively via DashboardWaypoints. No re-export
shim per the orchestrator resolution; no behavior change.
